clean endpoints
