+++ src/modules/bot/index.ts
@@ -1,11 +1,1 @@
-import axios from 'axios';
-
-export const postDiscordMessage = async ({ text, endpoint }: { text: string; endpoint: string }): Promise<void> => {
-    await axios
-        .post(endpoint, {
-            username: 'Astar Faucet',
-            avatar_url: '',
-            content: text,
-        })
-        .catch((e) => console.error(e.message));
-};
+export * from './discordMsg';
